Restructure into installable src/phishintention package (pyproject + python -m)
Adopt the production 'src layout': move modules/, networks/, utils/ and the
configs.py/phishintention.py entry points into src/phishintention/ (as
config.py and pipeline.py), add pyproject.toml with a console entry point, and
rename unittest_code/ -> tests/.

- All intra-project imports now use the absolute 'phishintention.*' namespace
- Run via 'python -m phishintention' (pipeline gains a main() entry point)
- config.py resolves configs/ from the project root instead of __file__
- setup.sh/setup.bat install the package editable ('pip install -e .')
- CI (run command, coverage include/omit, test path), ruff and SonarCloud
  exclusions, README and .gitignore updated accordingly

Validated locally that the package and every submodule resolve via an editable
install (remaining import errors are only the absent torch/detectron2/cv2);
full functional validation runs in CI.
