+++ sonar-project.properties
@@ -1,7 +1,7 @@
 sonar.projectKey=miaoyuxun_PhishIntention
 sonar.organization=miaoyuxun
 sonar.qualitygate.skip=true
-sonar.exclusions=networks/ocr/**
+sonar.exclusions=src/phishintention/networks/ocr/**
 sonar.python.coverage.reportPaths=coverage.xml
 
 # This is the name and version displayed in the SonarCloud UI.
