+++ src/phishintention/__init__.py
@@ -1,0 +1,3 @@
+"""PhishIntention: vision-based phishing detection (USENIX Security '22)."""
+
+__version__ = "1.0.0"
