+++ src/phishintention/__main__.py
@@ -1,0 +1,4 @@
+from phishintention.pipeline import main
+
+if __name__ == "__main__":
+    main()
