+++ tests/test_all_steps.py
